refactor: seperate HomeUiState and viewmodel
