+++ app/src/main/java/ac/sbmax002/eye_on/ui/home/HomeUiState.kt
@@ -1,0 +1,11 @@
+package ac.sbmax002.eye_on.ui.home
+
+data class HomeUiState(
+    val isReady: Boolean = false,
+    val isMonitoring: Boolean = false,
+    val isFaceDetected: Boolean = false,
+    val monitoringStartTime: Long = 0L,
+    val lastSessionDuration: Long = 0L,
+    val drowsinessDetectionCount: Int = 0
+)
+
